requirements.txt
absl-py==0.7.0
astor==0.7.1
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
bert-serving-client==1.8.2
bert-serving-server==1.8.2
boto3==1.9.65
botocore==1.12.65
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
en-core-web-sm==2.0.0
-e git+https://github.com/emayfield/debate-server.git@75d60b7079f6705b57c53be816b7e501dbd6514d#egg=endpoints
ffmpeg-normalize==1.3.8
gast==0.2.2
GPUtil==1.4.0
grpcio==1.18.0
h5py==2.8.0
idna==2.7
isort==4.3.4
Jinja2==2.10
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.3
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
networkx==2.1
nltk==3.4
numpy==1.15.0
pandas==0.23.4
pbr==5.1.2
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
protobuf==3.6.1
py==1.7.0
pylint==2.1.1
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==18.0.0
regex==2017.11.9
requests==2.20.1
s3transfer==0.1.13
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
spacy==2.0.17
tensorboard==1.12.2
tensorflow==1.13.0rc2
tensorflow-estimator==1.13.0
termcolor==1.1.0
thinc==6.12.0
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
